requirements.txt
streamlit==1.50.0
openai==1.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
plotly==5.15.0
pandas==2.3.3
pydantic==2.5.0
python-docx==0.8.11
python-multipart==0.0.6
uuid6==2023.5.2
sqlalchemy==2.0.23
bcrypt==4.0.1
pyjwt==2.8.0
